requirements.txt
# Core Async + Web Framework
fastapi>=0.104.1
uvicorn>=0.24.0.post1
sse-starlette>=2.3.5
starlette>=0.46.2
aiohttp>=3.12.0
aiofiles>=23.2.1
websockets>=13.1

# OpenAI + LangChain (modular architecture)
openai>=1.3.3
langchain-core>=0.3.61
langchain-community>=0.3.17
langchain-openai>=0.3.6
langchain-text-splitters>=0.3.8
langgraph>=0.2.76
langgraph-cli>=0.2.10
langgraph-sdk>=0.1.70
langgraph-checkpoint>=2.0.26
langsmith>=0.3.42

# Optional model connectors
langchain-ollama>=0.3.3
litellm>=1.71.0
fpdf2>=2.7.8

# Data + Document Handling
unstructured>=0.13
unstructured-client>=0.35.0
python-docx>=1.1.0
htmldocx>=0.0.6
beautifulsoup4>=4.12.2
markdown>=3.5.1
markdown2>=2.5.3
lxml>=4.9.2
pymupdf>=1.23.6
pdfminer.six>=20221105
docopt>=0.6.2

# File, Network, and Utility Libraries
requests>=2.31.0
httpx>=0.28.1
httpx-sse>=0.4.0
httpx-aiohttp>=0.1.4
urllib3>=2.4.0
certifi>=2025.4.26
charset-normalizer>=3.4.2
python-dotenv>=1.0.0
filelock>=3.18.0
tqdm>=4.67.1
regex>=2024.11.6
tenacity>=9.1.2

# JSON / Validation / Serialization
pydantic>=2.5.1
pydantic-core>=2.33.2
pydantic-settings>=2.9.1
dataclasses-json>=0.6.7
jsonschema>=4.23.0
jsonschema-specifications>=2025.4.1
orjson>=3.10.18
typing-extensions>=4.13.2
typing-inspect>=0.9.0

# NLP + Text Processing
tiktoken>=0.7.0
rapidfuzz>=3.13.0
tokenizers>=0.21.1

# Databases + Vector Stores
sqlalchemy>=2.0.28
fsspec>=2025.5.1

# Search + Knowledge Tools
duckduckgo-search>=4.1.1
arxiv>=2.0.0
feedparser>=6.0.11

# Compression + Utils
brotli>=1.1.0
zstandard>=0.23.0
zopfli>=0.2.3.post1

# Misc. Infra + Logs
loguru>=0.7.2
colorama>=0.4.6
attrs>=25.3.0
nest-asyncio>=1.6.0
anyio>=4.9.0

# Optional: Model Context Protocol (MCP)
mcp>=1.0.0
langchain-mcp-adapters>=0.1.0

# Windows-specific utility
win32-setctime>=1.2.0

# Optional: Supabase
supabase>=2.0.0
PyJWT>=2.8.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.9

json_repair
json5
mistune
pydantic[email]

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
faker>=19.0.0